backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-generativeai==0.3.2
beautifulsoup4==4.12.2
python-dotenv==1.0.0
